refactor: delete add
jire: DCS-1980
