+++ .github/workflows/PublishWorkbookWorkflow.yml
@@ -1,5 +1,6 @@
 name: PublishWorkbook
 on:
+  workflow_dispatch:
   pull_request:
     branches:
       - main
